:sparkles: Add mutedText color to ThemeExt
